requirements.txt
flask
flask-cors
pandas
numpy
bcrypt
pyjwt
scikit-learn
mysql-connector-python